feat: single source of truth — CapabilityLevel re-exported from portcullis-core
The verified type IS the production type. portcullis now re-exports
CapabilityLevel from portcullis-core instead of defining its own copy.
One type, one source of truth, zero translation layers.

The serde feature flag on portcullis-core enables serialization support
when used by portcullis (production) while keeping portcullis-core
dependency-free when used by Aeneas (verification).
